model/gpsdO1/requirements.txt
tqdm
loguru
torchmetrics
faiss-cpu
adamw_bf16